Avoid floating point error accumulation in PeriodicController
This is the classic problem of `0.1 + 0.1 + 0.1 != 0.3`, but the periodic callback checker that was there assumes it's going to be exact, and checks exactly above and below. However, we can fix this by instead using `start + steps*dt` for calculating the current location. This is the change that happened internally to DifferentialEquations and thus it was more correct, but sadly that made it different from the calculation here which exposed the floating point accumulation problem. However, by doing the same thing `start + steps * dt` here, we can avoid accumulating floating point error and thus keep the accurate check. Another way to fix this is by things like `+eps(t)`, but this is probably nicer because it directly fixes the problem of accumulation in the first place.